#release 0.6.0 log demodulated response

+++ src/main/scala/icfpc/classified/game/Vector.scala
@@ -3,11 +3,11 @@ package icfpc.classified.game
 case class Vector(x: Int, y: Int) {
   def +(other: Vector): Vector = Vector(x + other.x, y + other.y)
 
-  def *(v: Double) = Vector((x * v).toLong, (y * v).toLong)
+  def *(v: Double) = Vector((x * v).toInt, (y * v).toInt)
 
   def norm: Vector =
     Vector(
-      x.max(-1L).min(1L),
-      y.max(-1L).min(1L)
+      x.max(-1).min(1),
+      y.max(-1).min(1)
     )
 }
